+++ src/main/java/com/woowa/woowakit/domain/model/Quantity.java
@@ -1,4 +1,4 @@
-package com.woowa.woowakit.domain.product.domain.product;
+package com.woowa.woowakit.domain.model;
 
 import java.util.Objects;
