+++ src/main/java/com/woowa/woowakit/domain/product/application/StockService.java
@@ -3,9 +3,9 @@ package com.woowa.woowakit.domain.product.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.woowa.woowakit.domain.model.Quantity;
 import com.woowa.woowakit.domain.product.domain.product.Product;
 import com.woowa.woowakit.domain.product.domain.product.ProductRepository;
-import com.woowa.woowakit.domain.product.domain.product.Quantity;
 import com.woowa.woowakit.domain.product.domain.stock.ExpiryDate;
 import com.woowa.woowakit.domain.product.domain.stock.Stock;
 import com.woowa.woowakit.domain.product.domain.stock.StockRepository;
