+++ src/test/java/com/woowa/woowakit/domain/product/domain/product/ProductTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import com.woowa.woowakit.domain.model.Quantity;
+
 @DisplayName("Product 도메인 엔티티 테스트")
 class ProductTest {
